Se volvió a colocar el decimal en lugar del double en la categoría.

+++ SGE.Entidades/Categoria.cs
@@ -3,17 +3,17 @@
     public class Categoria
     {
         public string Codigo { get; set; }
-        public double ConsumoMinimo { get; set; }
-        public double ConsumoMaximo { get; set; }
-        public double CostoFijo { get; set; }
-        public double CostoVariable { get; set; }
+        public decimal ConsumoMinimo { get; set; }
+        public decimal ConsumoMaximo { get; set; }
+        public decimal CostoFijo { get; set; }
+        public decimal CostoVariable { get; set; }
 
-        public bool ConsumoDentroDeLosLimites(double consumo)
+        public bool ConsumoDentroDeLosLimites(decimal consumo)
         {
             return this.ConsumoMinimo < consumo && consumo < this.ConsumoMaximo;
         }
 
-        public double CalcularFacturaMensual(double consumo)
+        public decimal CalcularFacturaMensual(decimal consumo)
         {
             return this.CostoFijo + (this.CostoVariable * consumo);
         }
